+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Convert the DocC source for _The Swift Programming Language_ book into an EPUB file.
 - The generated EPUB file follows the rendering style used by Apple Books for [TSPL publications up to Swift 5.7](https://books.apple.com/us/book/the-swift-programming-language-swift-5-7/id881256329) and retains all internal references and external links.
 - Add new `--output-cover-image` flag to save the generated cover image as a separate PNG file in the output directory when generating an EPUB file.
+- Add new `--cover-footer-line` option to include a custom line of text in the cover image footer when generating an EPUB file.
 
 ### Internal
 - Add `pillow` dependency for image processing during EPUB generation.